+++ blocks/footer/footer.js
@@ -1,4 +1,4 @@
-import { readBlockConfig, decorateIcons } from '../../scripts/scripts.js';
+import { readBlockConfig, decorateIcons } from '../../scripts/block-utils.js';
 
 /**
  * loads and decorates the footer
